+++ Manga/app/src/main/java/nhdphuong/com/manga/data/BookDataSource.kt
@@ -11,8 +11,8 @@ import java.util.*
  */
 interface BookDataSource {
     interface Remote {
-        suspend fun getBookByPage(page: Int): RemoteBook?
-        suspend fun getBookByPage(searchContent: String, page: Int): RemoteBook?
+        suspend fun getBookByPage(page: Long): RemoteBook?
+        suspend fun getBookByPage(searchContent: String, page: Long): RemoteBook?
         suspend fun getBookDetails(bookId: String): Book?
         suspend fun getRecommendBook(bookId: String): RecommendBook?
     }
